[+] ShortenExtension option. [+] Ctrl+C handler. [+] User has aborted conversion return code.
git-svn-id: svn://svn.code.sf.net/p/xps2img/code/trunk@471 f56d0dea-3dc4-4344-980a-dda9636af30f

+++ xps2img/CommandLine/CommandLine.cs
@@ -56,7 +56,7 @@ namespace Xps2Img.CommandLine
                                     #endif
                                     , exceptionHint));
 
-            return (int)(ex is ConversionException ? (ex as ConversionException).ReturnCode : ReturnCode.Failed);
+            return ex is ConversionException ? (ex as ConversionException).ReturnCode : ReturnCode.Failed;
         }
 
         #endregion
